requirements.txt
python-telegram-bot==20.*
requests==2.31.*
python-dotenv==1.0.*
PyGithub==2.1.*